disable http keepalives to avoid fd leaks
When an already established ssh session hangs, the Go HTTP client still
has a TCP-over-HTTP connection in its keep-alive pool. This will be
re-used for any new connection. However, this bad cached connection
seems to get never removed, resulting in the accumulation of more and
more http client connections which hardly timeout.

Disabling HTTP keep alives avoids this problem. The side effect is a
performance decrease in connection handling of about 7%, which seems
acceptible for now.