ssh: close dead connection after depooling
This fixes a possible race condition where a
valid client is in the pool, a single port connect
fails, the client is closed (but still in the pool).
Concurrently, another goroutine grabs that client
from the pool.
Shortly afterwards, the first goroutine closes the
client and essentially breaks the second connection
attempt.